+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3 (2023-04-02)
+
+- Cleaning headers when follow redirect is active.
+- Fixing test matrix.
+- Updating tests.
+
 ## 0.0.2 (2023-03-26)
 
 - CI Publishing.